Add default values to config

+++ pippi/config.py
@@ -3,12 +3,31 @@ import shutil
 
 default_path = os.path.dirname(os.path.abspath(__file__)) + '/default.config'
 
+def parse(config_file):
+    default_config = {}
+    execfile(default_path, default_config)
+    if not os.path.isfile(config_file):
+        return default_config
+
+    config = {}
+    execfile(config_file, config)
+
+    for key in default_config.iterkeys():
+        if key not in config:
+            config[key] = default_config[key]
+
+    return config
+
 def init_user_config():
     config_path = os.path.expanduser('~/pippi.config')
-    shutil.copy(default_path, config_path)
-    return execfile(default_path, {})
+    if not os.path.isfile(config_path):
+        shutil.copy(default_path, config_path)
 
-def config():
+    config = {}
+    execfile(config_path, config)
+    return config
+
+def init():
     """ Look for a config file and load into dict, otherwise load defaults
     """
     config = None
@@ -17,10 +36,10 @@ def config():
 
     # look for a local config file
     if os.path.isfile(user_config_path):
-        config = execfile(config_file, {})
+        config = parse(user_config_path)
 
     if os.path.isfile(local_config_path):
-        config = execfile(config_file, {})
+        config = parse(local_config_path)
 
     if config is None:
         config = init_user_config()
